Don’t import MDX files
The MDX plugin sets the default export to the file
content and this doesn’t play well with
react-static which dynamically loads content as
needed.

Instead, write (another) script, this time to
generate page_listing.js by reading the MDX files.
It’s a bit hacky but it’ll do for now.